wip: today's progress, mainly comments addressing and code refactoring
Do code refactoring for tests, and address some of colin's comments.
The plan currently was to use the conversion of sysuser entries
into actual files, and temporarily create a rpm-ostree command
to convert /etc/passwd or /etc/group into sysuser entries.

Those will be coming in a later commit for testing