phase_one: conversion partly done
per title, the conversion is partly done. Really rewriting a code is so much better than just
changing the code.

You cna change the code, but the only case that will be beneficial so far is that you will be refactoring the code. Otherwise, if it is a mess already, you just rewrite it